Split Profiles report parser into layered modules (prof-report-parser-layers)
Refactor slice A parser into cuppa/cpp/profiles_report/: inventory and scope
handling, Alliance Clang line parsing, and profile-keyed classifiers starting
with std::init. classify_rule(profile, message) no longer applies std::init
patterns to other profile names. cxx_profiles_report.py remains a thin import
facade.

Add examples/profiles-std-init-violations/ and std_init_golden.json fixtures
for spec-driven capture of diagnostic text before slice C HTML. Groundwork
for #184.
