+++ cuppa/cpp/profiles_report/constants.py
@@ -1,0 +1,8 @@
+#          Copyright Jamie Allsop 2026-2026
+# Distributed under the Boost Software License, Version 1.0.
+#    (See accompanying file LICENSE_1_0.txt or copy at
+#          http://www.boost.org/LICENSE_1_0.txt)
+
+UNCLASSIFIED_RULE_ID = '_unclassified'
+
+DEFAULT_COMPILER = 'clang'
