+++ design/plans/cxx-profiles-report.md
@@ -614,8 +614,8 @@ Target cycle: **1.8.0** for **`prof-report-parser` … `prof-report-manifest`** 
 |-------|--------|
 | Plan | **This document** (`prof-report-collector` spawn scope settled) |
 | A — `prof-report-parser` | **Done** — merged [#190](https://github.com/ja11sop/cuppa/pull/190) |
-| B — `prof-report-collector` | **On branch** — PR after slice B lands ([#184](https://github.com/ja11sop/cuppa/issues/184)) |
-| B½ — `prof-report-parser-layers` | **Planned** — see §Parser layering follow-on (before slice C) |
+| B — `prof-report-collector` | **Done** — merged [#191](https://github.com/ja11sop/cuppa/pull/191) |
+| B½ — `prof-report-parser-layers` | **In progress** — branch `prof-report-parser-layers` |
 | C — `prof-report-html` | Not started |
 | D — `prof-report-manifest` | Not started |
 | E — `prof-report-method` | Deferred |
@@ -640,6 +640,10 @@ assumes ``std::init`` prose even though the **profile name** is parsed from
 
 Land **before slice C HTML** so doc links and rule sections live in profile modules, not in the
 generic report builder.
+
+**Status (2026-08-11):** landed on branch `prof-report-parser-layers` — package layout below,
+profile-keyed ``classify_rule``, golden fixtures, and ``examples/profiles-std-init-violations/``.
+Extend classifiers when new Clang message shapes are captured from the example or live trees.
 
 ### Target layout
