+++ examples/profiles-std-init-violations/sconscript
@@ -1,0 +1,3 @@
+Import( 'env' )
+
+env.Build( 'violations', 'violations.cpp' )
