+++ examples/profiles-std-init-violations/sconstruct
@@ -1,0 +1,5 @@
+import cuppa
+
+cuppa.run(
+    default_variants = [ 'dbg' ],
+)
